Restore consul presence testing to main_test.go
- Remove unused profiler port config so that it compiles under stricter modern Go
- Return "consul not configured" Context
- Return "consul" describe block (not sure why this isn't another
context, but this is the way it was)
- Introduce 'consulapi' custom package name to simplify 'api' package
name conflict

Note that getting these tests to run requires the 'consul' binary to be
found on the PATH.  It also must be version '0.9.4' or before.

This corresponds to SHA cc710aed in github.com/hashicorp/consul.  Other
versions can be found by checking the history of
consul/version/version.go (they don't appear to use tags or branches,
because they hate good things).

[#163434100]

Signed-off-by: Steve Taylor <staylor@pivotal.io>